% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzip_knn.R
\name{get_gzip_complexity}
\alias{get_gzip_complexity}
\title{Gzip Complexity}
\usage{
get_gzip_complexity(x)
}
\arguments{
\item{x}{single data instance}
}
\value{
the complexity of the data instance
}
\description{
The complexity of a data instance is calculated as the length of the
compressed representation of the serialized object.
}
