% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzip_knn.R
\name{gzip_dist}
\alias{gzip_dist}
\title{Gzip Distance}
\usage{
gzip_dist(ref, x)
}
\arguments{
\item{ref}{reference data instance}

\item{x}{single data instance to classify}
}
\value{
the distance between the two data instances
}
\description{
The distance is calculated as the difference between the complexity of
the reference and the complexity of the test point, divided by the
maximum complexity between the two.
}
