% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gzip_knn.R
\name{gzip_knn}
\alias{gzip_knn}
\title{Gzip KNN}
\usage{
gzip_knn(x, model, k = 3)
}
\arguments{
\item{x}{single data instance to classify}

\item{model}{gzipr model}

\item{k}{number of neighbors to consider}
}
\value{
the class of the majority of the \code{k} nearest neighbors
}
\description{
For each of the \code{k} nearest neighbors, the distance is calculated as
the difference between the complexity of the reference and the
complexity of the test point, divided by the maximum complexity
between the two.
}
